1.2.1 final (release).

+++ menu/menu_core.c
@@ -150,7 +150,7 @@ void menu_core_catchButtonPress(void)
 		}
 		else if (menu_item[item_highlighted].type == 2)
 		{
-			if (menu_item[item_highlighted].extra_val != 0 && menu_item[item_highlighted].float_val == menu_item[item_highlighted].extra_val)
+			if (menu_item[item_highlighted].extra_val != 0 && menu_item[item_highlighted].float_val == (float) menu_item[item_highlighted].extra_val)
 				menu_item[item_highlighted].float_val = 0.0;
 			else
 			{
@@ -164,7 +164,6 @@ void menu_core_catchButtonPress(void)
 			
 			PLAY_AUDIO_EVENT("FRONTEND_MENU_MP_SERVER_OPTION_CHANGE");
 		}
-
 	}
 
 	if (menu_forward_pressed() && press_counter_timesby < 2 && !inError)
